@@ -1,13 +1,13 @@
 name: dwds
 # Every time this changes you need to run `dart run build_runner build`.
-version: 16.0.0-dev
+version: 16.0.0
 description: >-
   A service that proxies between the Chrome debug protocol and the Dart VM
   service protocol.
 repository: https://github.com/dart-lang/webdev/tree/master/dwds
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   async: ^2.9.0
