@@ -4,7 +4,7 @@ description: A web app example for webdev CLI.
 publish_to: none
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dev_dependencies:
   build_runner: ^2.0.0
