@@ -12,7 +12,7 @@ description:
 # and build_web_compilers constraint should match those defined
 # in pubspec.dart.
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dev_dependencies:
   build_runner: '>=1.6.2 <3.0.0'
