@@ -2,7 +2,7 @@ name: _webdev_smoke
 description: A test fixture for webdev testing with sound support.
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dev_dependencies:
   build_runner: '>=1.6.2 <3.0.0'
