@@ -9,7 +9,7 @@ description: >-
 repository: https://github.com/dart-lang/webdev/tree/master/webdev
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   args: ^2.3.1
