@@ -2,4 +2,5 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/readers/asset_reader.dart' show AssetReader, UrlEncoder;
+export 'src/readers/asset_reader.dart'
+    show AssetReader, UrlEncoder, PackageUriMapper, stripLeadingSlashes;
