@@ -75,7 +75,7 @@ class BuildRunnerRequireStrategyProvider {
   Future<String?> _moduleForServerPath(
       MetadataProvider metadataProvider, String serverPath) async {
     final modulePathToModule = await metadataProvider.modulePathToModule;
-    final relativePath = relativizePath(serverPath);
+    final relativePath = stripLeadingSlashes(serverPath);
     for (var e in modulePathToModule.entries) {
       if (stripTopLevelDirectory(e.key) == relativePath) {
         return e.value;
@@ -97,10 +97,14 @@ class BuildRunnerRequireStrategyProvider {
     return stripTopLevelDirectory(path);
   }
 
-  String? _serverPathForAppUri(String appUri) {
-    if (appUri.startsWith('org-dartlang-app:')) {
+  String? _serverPathForAppUri(String appUrl) {
+    final appUri = Uri.parse(appUrl);
+    if (appUri.isScheme('org-dartlang-app')) {
       // We skip the root from which we are serving.
-      return Uri.parse(appUri).pathSegments.skip(1).join('/');
+      return appUri.pathSegments.skip(1).join('/');
+    }
+    if (appUri.isScheme('package')) {
+      return '/packages/${appUri.path}';
     }
     return null;
   }
