@@ -1,2 +1,2 @@
 // Generated code. Do not modify.
-const packageVersion = '16.0.0-dev';
+const packageVersion = '16.0.0';
