@@ -3,12 +3,13 @@
 // found in the LICENSE file.
 
 // Note: this is a copy from flutter tools, updated to work with dwds tests,
-// and some functionality remioved (does not support hot reload yet)
+// and some functionality removed (does not support hot reload yet)
 
 import 'dart:async';
 
 import 'package:dwds/asset_reader.dart';
 import 'package:dwds/expression_compiler.dart';
+import 'package:file/file.dart';
 import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
 
@@ -29,6 +30,7 @@ class ResidentWebRunner {
       this.urlTunneler,
       this.projectDirectory,
       this.packageConfigFile,
+      this.packageUriMapper,
       this.fileSystemRoots,
       this.fileSystemScheme,
       this.outputPath,
@@ -38,6 +40,7 @@ class ResidentWebRunner {
       dartSdkPath,
       projectDirectory: projectDirectory,
       packageConfigFile: packageConfigFile,
+      useDebuggerModuleNames: packageUriMapper.useDebuggerModuleNames,
       platformDill:
           soundNullSafety ? '$platformDillSound' : '$platformDillUnsound',
       fileSystemRoots: fileSystemRoots,
@@ -51,6 +54,7 @@ class ResidentWebRunner {
   final Uri mainUri;
   final Uri projectDirectory;
   final Uri packageConfigFile;
+  final PackageUriMapper packageUriMapper;
   final String outputPath;
   final List<Uri> fileSystemRoots;
   final String fileSystemScheme;
@@ -62,13 +66,14 @@ class ResidentWebRunner {
   late Uri uri;
   late Iterable<String> modules;
 
-  Future<int> run(String? hostname, int port, String index) async {
+  Future<int> run(
+      FileSystem fileSystem, String? hostname, int port, String index) async {
     devFS = WebDevFS(
       fileSystem: fileSystem,
       hostname: hostname ?? 'localhost',
       port: port,
       projectDirectory: projectDirectory,
-      packageConfigFile: packageConfigFile,
+      packageUriMapper: packageUriMapper,
       index: index,
       urlTunneler: urlTunneler,
       soundNullSafety: soundNullSafety,
