@@ -163,7 +163,7 @@ Future<List<PackageExceptionDetails>> _validateBuildDaemonVersion(
 }
 
 final buildRunnerConstraint = VersionConstraint.parse('>=1.6.2 <3.0.0');
-final buildWebCompilersContraint = VersionConstraint.parse('>=2.12.0 <4.0.0');
+final buildWebCompilersConstraint = VersionConstraint.parse('>=2.12.0 <4.0.0');
 
 // Note the minimum versions should never be dev versions as users will not
 // get them by default.
@@ -178,7 +178,7 @@ Future<void> checkPubspecLock(PubspecLock pubspecLock,
 
   if (requireBuildWebCompilers) {
     issues.addAll(pubspecLock.checkPackage(
-        'build_web_compilers', buildWebCompilersContraint));
+        'build_web_compilers', buildWebCompilersConstraint));
   }
 
   if (buildRunnerIssues.isEmpty) {
